Regenerate 4096-bit RSA keys
The command `service sshd keygen` does not take into consideration the
option `-b`, which specifies the number of bits in the key to create.
For RSA, 2048-bit keys are generated by default on FreeBSD 12.2 and 13.0
(OpenSSH 7.9), and 3072-bit on FreeBSD 14.0 (OpenSSH 8.7).

Fix it by using the full command, just like the other guides.